pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wagtail-translate"
authors = [{name = "Coen van der Kamp", email = "coen@fourdigits.nl"}]
description = "Use Wagtail Translate to machine translate your Wagtail contents."
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 5",
    "Framework :: Wagtail :: 6",
]
requires-python = ">=3.8"
dynamic = ["version"]
dependencies = [
    "Django>=4.2",
    "Wagtail>=5.2"
]
[project.optional-dependencies]
testing = [
    "dj-database-url==2.1.0",
    "pre-commit==3.4.0",
    "pytest==8.1.1",
    "pytest-cov==5.0.0",
    "pytest-django==4.8.0",
    "pytest-factoryboy==2.7.0",
]
ci = [
    "tox==4.11.3",
    "tox-gh-actions==3.1.3",
]

[project.urls]
Home = "https://github.com/allcaps/wagtail-translate"

[tool.flit.module]
name = "wagtail_translate"

[tool.flit.sdist]
exclude = [
    "manage.py",
    "*.yaml",
    "ruff.toml",
    ".*",
    "*.js",
    "*.json",
    "*.ini",
    "*.yml"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.testproject.settings"
